feat(MenuEditor): 添加自定义事件标签组件，简化自定义事件代码输入流程
